pyproject.toml
[project]
name = "unival"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandera[polars]>=0.24.0",
    "polars>=1.29.0",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
]
